refactor: changed name of SendToWalletButton to ShowSendToWalletButton and ofSendToWallet to SendToWalletButton
