--- packages/widget/src/components/SendToWallet/index.ts
@@ -1,2 +1,0 @@
-export * from './SendToWallet';
-export * from './SendToWalletButton';
