+++ packages/widget/src/components/SendToWalletButton/ShowSendToWalletButton.tsx
@@ -7,7 +7,7 @@ import { FormKey, useWallet, useWidgetConfig } from '../../providers';
 import { useSendToWalletStore, useSettings } from '../../stores';
 import { DisabledUI, HiddenUI, RequiredUI, WidgetEvent } from '../../types';
 
-export const SendToWalletButton: React.FC = () => {
+export const ShowSendToWalletButton: React.FC = () => {
   const { t } = useTranslation();
   const { setValue } = useFormContext();
   const emitter = useWidgetEvents();
