+++ packages/widget/src/components/SendToWalletButton/index.ts
@@ -1,0 +1,2 @@
+export * from './SendToWalletButton';
+export * from './ShowSendToWalletButton';
